+++ Source/Modules/Catalog/Math/Gapotchenko.FX.Math.Intervals/IntervalEqualityComparer`1.Generic.cs
@@ -11,15 +11,28 @@ partial class IntervalEqualityComparer<T>
 {
     internal sealed class GenericComparer(IEqualityComparer<T>? comparer) : IntervalEqualityComparer<T>
     {
-        public override bool Equals(IInterval<T>? x, IInterval<T>? y) =>
-            ReferenceEquals(x, y) ||
-            x is not null && y is not null &&
-            x.From.Equals(y.From, comparer) &&
-            x.To.Equals(y.To, comparer);
+        public override bool Equals(IInterval<T>? x, IInterval<T>? y)
+        {
+            return
+                ReferenceEquals(x, y) ||
+                x is not null && y is not null &&
+                (x.IsEmpty && y.IsEmpty ||
+                x.From.Equals(y.From, comparer) &&
+                x.To.Equals(y.To, comparer));
+        }
 
-        public override int GetHashCode(IInterval<T> obj) =>
-            HashCode.Combine(
-                obj.From.GetHashCode(comparer),
-                obj.To.GetHashCode(comparer));
+        public override int GetHashCode(IInterval<T> obj)
+        {
+            if (obj.IsEmpty)
+            {
+                return 0;
+            }
+            else
+            {
+                return HashCode.Combine(
+                    obj.From.GetHashCode(comparer),
+                    obj.To.GetHashCode(comparer));
+            }
+        }
     }
 }